// multi-step-form/src/types.ts
export interface IFormState {
    [key: string]: any;
    step: number;
    name: IInput;
    email: IInput;
    tel: IInput;
    plan: IPlanOption;
    period: TPeriod;
    addOns: IAddOn[];
    isCompleted: boolean;
}

export interface IPlanOption {
    name: TPlan;
    price: number;
    monthsFree: number;
}

export interface IAddOn {
    heading: string;
    description: string;
    price: number;
}

export interface IInput {
    value: string;
    error: string;
}

export type TPlan = "Arcade" | "Advanced" | "Pro";
export type TPeriod = "Monthly" | "Yearly";